fix: rekognition

+++ pages/api/checkForInappropriateContent.js
@@ -1,9 +1,8 @@
-import { Amplify, withSSRContext } from "aws-amplify";
+import { withSSRContext } from "aws-amplify";
 import { Rekognition } from "@aws-sdk/client-rekognition";
 import aws from "aws-sdk";
 
 const { Credentials } = withSSRContext({});
-
 const credentials = await Credentials.get();
 //Put this inside the function
 const rekognition = new Rekognition({
@@ -33,31 +32,25 @@ export const checkForInappropriateContent = async (key) => {
   try {
     const moderationLabels = await rekognition.detectModerationLabels(params);
 
-    if (moderationLabels.ModerationLabels.length > 0) {
-    
-      if (moderationLabels.ModerationLabels[1].Name === "Explicit Nudity") {
-        // const command = new DeleteObjectCommand(params);
-        // const response = await s3Client.send(command);
-        // return response
-        const response = {
-          inappropriate: true,
-          moderationLabels: {
-            name: moderationLabels.ModerationLabels[0].Name,
-            parentName: moderationLabels.ModerationLabels[0].ParentName,
-            confidence: moderationLabels.ModerationLabels[0].Confidence,
-          },
-        };
-        bucket.deleteObject({ Key: key, Bucket: bucketName }, (err) => {
-          console.log({ err });
-        });
-        return response;
-      } else {
-        const response = {
-          inappropriate: false,
-        };
-
-        return response;
-      }
+    if (
+      moderationLabels.ModerationLabels.length > 0 &&
+      moderationLabels.ModerationLabels[1].Name === "Explicit Nudity"
+    ) {
+      // const command = new DeleteObjectCommand(params);
+      // const response = await s3Client.send(command);
+      // return response
+      const response = {
+        inappropriate: true,
+        moderationLabels: {
+          name: moderationLabels.ModerationLabels[0].Name,
+          parentName: moderationLabels.ModerationLabels[0].ParentName,
+          confidence: moderationLabels.ModerationLabels[0].Confidence,
+        },
+      };
+      bucket.deleteObject({ Key: key, Bucket: bucketName }, (err) => {
+        console.log({ err });
+      });
+      return response;
     } else {
       const response = {
         inappropriate: false,
